dpkg: Add control over removal of built packages
In almost all cases, the packages built by a recipe wasn't used yet and
will only be requested down the dependency chain. Then it is fine to
purge it from the buildchroots during partial rebuilds.

There are corner cases, though, when the packages will never be used in
the buildchroots, only in the target filesystem. That can affect core
packages rebuilt for the target but also used in their upstream variant
in the buildchroots. While buildchroot installation can be controlled
via preferences, purging can't this way. Add a variable that allows to
control it without having to overwrite deb_clean.

+++ meta/classes/dpkg-base.bbclass
@@ -154,6 +154,8 @@ python do_dpkg_build() {
 
 addtask dpkg_build before do_build
 
+KEEP_INSTALLED_ON_CLEAN ?= "0"
+
 CLEANFUNCS += "deb_clean"
 
 deb_clean() {
@@ -162,6 +164,9 @@ deb_clean() {
         for d in ${DEBS}; do
             repo_del_package "${REPO_ISAR_DIR}"/"${DISTRO}" \
                 "${REPO_ISAR_DB_DIR}"/"${DISTRO}" "${DEBDISTRONAME}" "${d}"
+            if [ "${KEEP_INSTALLED_ON_CLEAN}" = "1" ]; then
+                continue;
+            fi
             package=$(basename "${d}")
             package_remove="/usr/bin/apt-get remove -y ${package%%_*}"
             sudo -E chroot ${BUILDCHROOT_DIR} ${package_remove} || true
